backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
boto3==1.29.7
python-dotenv==1.0.0
drf-spectacular==0.26.5
Pillow==10.1.0
